# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/constellation": {
            "get": {
                "description": "Возвращает список всех не удаленных доставок",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Созвездие"
                ],
                "summary": "Получение списка доставок",
                "parameters": [
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Номер рейса",
                        "name": "searchFlightNumber",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Начало даты формирования",
                        "name": "startFormationDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Конец даты формирования",
                        "name": "endFormationDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Статус созвездия",
                        "name": "constellationStatus",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список доставок",
                        "schema": {
                            "$ref": "#/definitions/model.ConstellationRequest"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ConstellationRequest"
                        }
                    }
                }
            }
        },
        "/constellation/{id}": {
            "get": {
                "description": "Возвращает информацию о доставке по её идентификатору",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Созвездие"
                ],
                "summary": "Получение созвездия по идентификатору",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор созвездия",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о доставке",
                        "schema": {
                            "$ref": "#/definitions/model.ConstellationGetResponse"
                        }
                    },
                    "400": {
                        "description": "Недопустимый идентификатор созвездия",
                        "schema": {
                            "$ref": "#/definitions/model.ConstellationGetResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ConstellationGetResponse"
                        }
                    }
                }
            }
        },
        "/constellation/{id}/delete": {
            "delete": {
                "description": "Удаляет доставку по её идентификатору",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Созвездие"
                ],
                "summary": "Удаление созвездия",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор созвездия",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Номер рейса",
                        "name": "searchFlightNumber",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Начало даты формирования",
                        "name": "startFormationDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Конец даты формирования",
                        "name": "endFormationDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Статус созвездия",
                        "name": "constellationStatus",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список планетаей",
                        "schema": {
                            "$ref": "#/definitions/model.ConstellationRequest"
                        }
                    },
                    "400": {
                        "description": "Недопустимый идентификатор созвездия",
                        "schema": {
                            "$ref": "#/definitions/model.ConstellationRequest"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ConstellationRequest"
                        }
                    }
                }
            }
        },
        "/constellation/{id}/status": {
            "put": {
                "description": "Обновляет статус созвездия для модератора по идентификатору созвездия",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Созвездие"
                ],
                "summary": "Обновление статуса созвездия для модератора",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор созвездия",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новый статус созвездия",
                        "name": "constellationStatus",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ConstellationUpdateStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о доставке",
                        "schema": {
                            "$ref": "#/definitions/model.ConstellationGetResponse"
                        }
                    },
                    "400": {
                        "description": "Недопустимый идентификатор созвездия или ошибка чтения JSON объекта",
                        "schema": {
                            "$ref": "#/definitions/model.ConstellationGetResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ConstellationGetResponse"
                        }
                    }
                }
            }
        },
        "/constellation/{id}/update": {
            "put": {
                "description": "Обновляет номер рейса для созвездия по её идентификатору",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Созвездие"
                ],
                "summary": "Обновление номера рейса созвездия",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор созвездия",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новый номер рейса",
                        "name": "flightNumber",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ConstellationUpdateFlightNumberRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о доставке",
                        "schema": {
                            "$ref": "#/definitions/model.ConstellationGetResponse"
                        }
                    },
                    "400": {
                        "description": "Недопустимый идентификатор созвездия или ошибка чтения JSON объекта",
                        "schema": {
                            "$ref": "#/definitions/model.ConstellationGetResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ConstellationGetResponse"
                        }
                    }
                }
            }
        },
        "/constellation/{id}/user": {
            "put": {
                "description": "Обновляет статус созвездия для пользователя по идентификатору созвездия",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Созвездие"
                ],
                "summary": "Обновление статуса созвездия для пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор созвездия",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о доставке",
                        "schema": {
                            "$ref": "#/definitions/model.ConstellationGetResponse"
                        }
                    },
                    "400": {
                        "description": "Недопустимый идентификатор созвездия",
                        "schema": {
                            "$ref": "#/definitions/model.ConstellationGetResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ConstellationGetResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Planet": {
            "type": "object",
            "properties": {
                "airline": {
                    "type": "string",
                    "example": "AirlineX"
                },
                "planet_code": {
                    "type": "string",
                    "example": "ABC123"
                },
                "planet_id": {
                    "type": "integer"
                },
                "planet_status": {
                    "type": "string",
                    "example": "checked"
                },
                "planet_type": {
                    "type": "string",
                    "example": "suitcase"
                },
                "owner_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "pasport_details": {
                    "type": "string",
                    "example": "123456789"
                },
                "photo": {
                    "type": "string",
                    "example": "http://example.com/planet.jpg"
                },
                "size": {
                    "type": "string",
                    "example": "large"
                },
                "weight": {
                    "type": "number",
                    "example": 23.5
                }
            }
        },
...
        "model.ConstellationGetResponse": {
            "type": "object",
            "properties": {
                "planets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Planet"
                    }
                },
                "confirmation_date": {
                    "type": "string"
                },
                "creation_date": {
                    "type": "string"
                },
                "constellation_id": {
                    "type": "integer"
                },
                "constellation_status": {
                    "type": "string"
                },
                "flight_number": {
                    "type": "string"
                },
                "formation_date": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                }
            }
        },
        "model.ConstellationRequest": {
            "type": "object",
            "properties": {
                "confirmation_date": {
                    "type": "string"
                },
                "creation_date": {
                    "type": "string"
                },
                "constellation_id": {
                    "type": "integer"
                },
                "constellation_status": {
                    "type": "string"
                },
                "flight_number": {
                    "type": "string"
                },
                "formation_date": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                }
            }
        },
        "model.ConstellationUpdateFlightNumberRequest": {
            "type": "object",
            "properties": {
                "flight_number": {
                    "type": "string"
                }
            }
        },
        "model.ConstellationUpdateStatusRequest": {
            "type": "object",
            "properties": {
                "constellation_status": {
                    "type": "string"
                }
            }
        },
...
    }
}